Move validate into dcpy.lifecyle.package
